+++ src/Feliz/React/React.fs
@@ -629,6 +629,41 @@ useLayoutEffect(() => {
     [<ImportMember("react")>]
     static member inline useImperativeHandle(ref: IRefValue<'t>, createHandle: unit -> 't, ?dependencies: obj[]) = jsNative
 
+    /// <summary>
+    /// Subscribes to a data source, and returns the current value from it.
+    /// </summary>
+    /// <param name='subscribe'>A function that sets up a subscription to the external data source. It receives a callback to be called when the data source changes.</param>
+    /// <param name='getSnapshot'>A function that returns the current value of the external data source.</param>
+    /// <returns>The current value from the external data source.</returns>
+    [<ImportMember("react")>]
+    static member inline useSyncExternalStore(subscribe: Func<(unit -> unit),(unit -> unit)> , getSnapshot: unit -> 'T, ?getServerSnapshot: unit -> 'T): 'T = jsNative
+
+    /// <summary>
+    /// Subscribes to a data source, and returns the current value from it.
+    /// </summary>
+    /// <param name='subscribe'>A function that sets up a subscription to the external data source. It receives a callback to be called when the data source changes.</param>
+    /// <param name='getSnapshot'>A function that returns the current value of the external data source.</param>
+    /// <returns>The current value from the external data source.</returns>
+    static member inline useSyncExternalStore(subscribe: (unit -> unit) -> (unit -> unit), getSnapshot: unit -> 'T, ?getServerSnapshot: unit -> 'T): 'T =
+        React.useSyncExternalStore( Func<_,_> subscribe, getSnapshot, ?getServerSnapshot = getServerSnapshot)
+
+    /// <summary>
+    /// Subscribes to a data source, and returns the current value from it.
+    /// </summary>
+    /// <param name='subscribe'>A function that sets up a subscription to the external data source. It receives a callback to be called when the data source changes.</param>
+    /// <param name='getSnapshot'>A function that returns the current value of the external data source.</param>
+    /// <returns>The current value from the external data source.</returns>
+    static member inline useSyncExternalStore(subscribe: (unit -> unit) -> #IDisposable, getSnapshot: unit -> 'T, ?getServerSnapshot: unit -> 'T): 'T =
+        React.useSyncExternalStore( Func<_,_> 
+            (fun (callback) ->
+                let disp = subscribe(callback)
+                fun () -> disp.Dispose()
+            ), 
+            getSnapshot,
+            ?getServerSnapshot = getServerSnapshot
+        )
+
+
 // This extensions module is required to help f# compiler understand overloads. 
 // Without this, for me the compiler was unable to resolve e.g. `useState` overload between `'t` and `unit -> 't`
 [<AutoOpen>]